+++ CHANGELOG.md
@@ -25,6 +25,7 @@
   - Update gaia to version 4.2.0 for e2e tests on CI ([#809])
   - Add `start-multi` command to relay on all paths defined in the configuration ([#748])
   - Add option to specify which events to listen for in `listen` command ([#550])
+  - Add option to customise receiver address for `ft-transfer` command ([#806])
   - Add `keys restore` command to import a signing key from its mnemonic ([#813])
 
 ### IMPROVEMENTS
@@ -106,8 +107,9 @@
 [#793]: https://github.com/informalsystems/ibc-rs/pull/793
 [#798]: https://github.com/informalsystems/ibc-rs/issues/798
 [#801]: https://github.com/informalsystems/ibc-rs/issues/801
-[#801]: https://github.com/informalsystems/ibc-rs/issues/809
 [#805]: https://github.com/informalsystems/ibc-rs/issues/805
+[#806]: https://github.com/informalsystems/ibc-rs/issues/806
+[#809]: https://github.com/informalsystems/ibc-rs/issues/809
 
 
 ## v0.1.1